DOC: moding Sphinx files into doc/
